Day 20 parts 1 and 2
I cheated for part 2 slightly, by looking at a graph of the nodes.
That was enough to see that the penultimate module was an inverter,
therefore I could calculate the cycle length of each of its inputs
separately.

My solution for part b also assumes that the first time we see a
false it marks a complete cycle. I didn't confirm that every
subsequent cycle is the same length.